Fixed Future Data Bug
When beginning to build the ai into the game, it became apparent that the training data was setup wrong. X was given information about the current turn AFTER the actions had already been selected, and so the real data it should be predicting on is from the previous turn.

With the new data the network still appeared to train, although ended at a higher loss around 3.5 instead of 2.5.